Time: 19 ms (6.31%), Space: 8.5 MB (88.01%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    long long mySqrt(long long x) {
-        long long ans = 1;
+    long mySqrt(long x) {
+        long ans = 1;
         if(x == 1 || x == 0){
             return x;
         }
